Added collision detection for upper and lower
boundary and get left and right x and get top and bottom y for
boundaries and pipes and heli

+++ src/heliController.h
@@ -26,10 +26,10 @@ public:
 	void draw();
 	void changeDirection();
 	int velocity( heliController & other);
-	int getRightX();
-	int getLeftX();
-	int getTopY();
-	int getBottomY();
+	float getRightX();
+	float getLeftX();
+	float getTopY();
+	float getBottomY();
 	void setisfalling(bool);
 
 };
